requirements.txt
python-telegram-bot==13.15
openpyxl==3.1.2
matplotlib==3.10.6
